src/requirements.txt
coloredlogs
pyyaml
celery
uvicorn
flower
redis
fastapi
python-dotenv
sqlalchemy
mysql-connector-python
